move footer and navbar to component folder
